Time: 29 ms (94.58%) | Memory: 102.6 MB (15.88%) - LeetSync

+++ 155-min-stack/min-stack.java
@@ -29,7 +29,7 @@ class MinStack {
         long x = st.pop();
         if(x < mini)
         {
-            mini = 2L * (int)mini - x;
+            mini = 2L * mini - x;
         }
     }
